# docs/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"APIResponse_null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"anyOf": [],
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"SeatingsList": {
				"description": "The response shape for seating creation and retrieval.",
				"properties": {
					"seatings": {
						"items": {
							"$ref": "#/components/schemas/SeatingResponseItem"
						},
						"type": "array"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
...
			"SeatingResponse": {
				"description": "The response shape for seating creation and retrieval.",
				"properties": {
					"seating": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/EventSeatMapping"
							},
							{
								"$ref": "#/components/schemas/SeatingsList"
							}
						],
						"description": "The base event seating mapping metadata."
					},
					"arrangement": {
						"anyOf": [
							{
								"$ref": "#/components/schemas/SectionAndRowsSeating"
							},
							{
								"$ref": "#/components/schemas/TableAndChairsSeating"
							},
							{
								"$ref": "#/components/schemas/MixedSeating"
							}
						],
						"description": "A detailed seating arrangement definition.\r\n\r\nCan be one of:\r\n- Section and rows seating\r\n- Table and chairs seating\r\n- Mixed seating"
					}
				},
				"type": "object",
				"additionalProperties": false
			},
			"APIResponse_SeatingResponse-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/SeatingResponse"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateSeatingInputBody": {
				"properties": {
					"seatingType": {
						"type": "string",
						"enum": [
							"section_n_rows",
							"table_n_chairs",
							"mixed",
							"individual"
						]
					},
					"totalCapacity": {
						"type": "number",
						"format": "double"
					},
					"availableCapacity": {
						"type": "number",
						"format": "double"
					},
					"noOfSections": {
						"type": "number",
						"format": "double"
					},
					"noOfRowsPerSection": {
						"type": "number",
						"format": "double"
					},
					"noOfChairsPerSection": {
						"type": "number",
						"format": "double"
					},
					"noOfTables": {
						"type": "number",
						"format": "double"
					},
					"noOfChairsPerTable": {
						"type": "number",
						"format": "double"
					},
					"totalChairTable": {
						"type": "number",
						"format": "double"
					},
					"noOfChairsPerRows": {
						"type": "number",
						"format": "double"
					},
					"totalSectionRow": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"seatingType",
					"totalCapacity",
					"availableCapacity"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UpdateSeatingInputBody": {
				"properties": {
					"seatingType": {
						"type": "string",
						"enum": [
							"section_n_rows",
							"table_n_chairs",
							"mixed",
							"individual"
						]
					},
					"totalCapacity": {
						"type": "number",
						"format": "double"
					},
					"availableCapacity": {
						"type": "number",
						"format": "double"
					},
					"noOfSections": {
						"type": "number",
						"format": "double"
					},
					"noOfRowsPerSection": {
						"type": "number",
						"format": "double"
					},
					"noOfTables": {
						"type": "number",
						"format": "double"
					},
					"noOfChairsPerTable": {
						"type": "number",
						"format": "double"
					},
					"totalChairTable": {
						"type": "number",
						"format": "double"
					},
					"noOfChairsPerRows": {
						"type": "number",
						"format": "double"
					},
					"totalSectionRow": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"seatingType"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"EventOnlyDTO": {
				"properties": {
					"id": {
						"type": "string"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"status": {
						"type": "string",
						"enum": [
							"completed",
							"published",
							"cancelled",
							"ongoing",
							"draft",
							null
						],
						"nullable": true
					},
					"organiserID": {
						"type": "string",
						"nullable": true
					},
					"eventName": {
						"type": "string"
					},
					"eventCategory": {
						"type": "string"
					},
					"eventDescription": {
						"type": "string"
					},
					"eventBanner": {
						"type": "string"
					},
					"publishedTimestamp": {
						"type": "string",
						"format": "date-time"
					},
					"isRecurringEvent": {
						"type": "boolean",
						"nullable": true
					},
					"isSingleEvent": {
						"type": "boolean",
						"nullable": true
					},
					"organizedBy": {
						"type": "string",
						"nullable": true
					},
					"public": {
						"type": "boolean",
						"nullable": true
					},
					"private": {
						"type": "boolean",
						"nullable": true
					},
					"publish": {
						"type": "boolean",
						"nullable": true
					},
					"schedule": {
						"type": "boolean",
						"nullable": true
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"status",
					"organiserID",
					"eventName",
					"eventCategory",
					"eventBanner",
					"publishedTimestamp",
					"isRecurringEvent",
					"isSingleEvent",
					"organizedBy",
					"public",
					"private",
					"publish",
					"schedule"
				],
				"type": "object",
				"additionalProperties": false
			},
			"APIResponse_EventOnlyDTO-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/EventOnlyDTO"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Omit_EventContactDTO.id-or-eventId-or-createdAt-or-updatedAt_": {
				"$ref": "#/components/schemas/Pick_EventContactDTO.Exclude_keyofEventContactDTO.id-or-eventId-or-createdAt-or-updatedAt__",
				"description": "Construct a type with the properties of T except for those in type K."
			},
			"AddEventContactDTO": {
				"$ref": "#/components/schemas/Omit_EventContactDTO.id-or-eventId-or-createdAt-or-updatedAt_"
			},
...
			"RecurringEventInput": {
				"$ref": "#/components/schemas/Omit_RecurringEvent.eventId-or-id-or-longitude-or-latitude-or-createdAt-or-updatedAt_"
			},
...
			"CreateEventDTO": {
				"properties": {
					"status": {
						"type": "string",
						"enum": [
							"completed",
							"published",
							"cancelled",
							"ongoing",
							"draft"
						]
					},
					"eventName": {
						"type": "string"
					},
					"eventCategory": {
						"type": "string"
					},
					"eventDescription": {
						"type": "string"
					},
					"eventBanner": {
						"type": "string"
					},
					"publishedTimestamp": {
						"type": "string",
						"format": "date-time"
					},
					"isRecurringEvent": {
						"type": "boolean"
					},
					"isSingleEvent": {
						"type": "boolean"
					},
					"organizedBy": {
						"type": "string"
					},
					"public": {
						"type": "boolean"
					},
					"private": {
						"type": "boolean"
					},
					"publish": {
						"type": "boolean"
					},
					"schedule": {
						"type": "boolean"
					},
					"contact": {
						"$ref": "#/components/schemas/AddEventContactDTO",
						"description": "Contact information for the event."
					},
					"tags": {
						"items": {
							"type": "string"
						},
						"type": "array",
						"description": "An array of tags associated with the event for categorization."
					},
					"recurrenceDetails": {
						"items": {
							"$ref": "#/components/schemas/RecurringEventInput"
						},
						"type": "array",
						"description": "An array of recurrence details for events that repeat on a schedule."
					}
				},
				"required": [
					"status",
					"eventName",
					"eventCategory",
					"eventDescription",
					"eventBanner",
					"publishedTimestamp",
					"isRecurringEvent",
					"isSingleEvent",
					"organizedBy",
					"public",
					"private",
					"publish",
					"schedule"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"APIResponse_PaginatedResult_Events_-or-null_": {
				"properties": {
					"success": {
						"type": "boolean"
					},
					"data": {
						"allOf": [
							{
								"$ref": "#/components/schemas/PaginatedResult_Events_"
							}
						],
						"nullable": true
					},
					"error": {
						"type": "string"
					},
					"message": {}
				},
				"required": [
					"success"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearer": {
				"type": "http",
				"name": "Authorization",
				"scheme": "bearer",
				"in": "header",
				"bearerFormat": "JWT",
				"description": "Enter your bearer token in the format \"Bearer {token}\""
			},
			"tsoa_auth": {
				"type": "oauth2",
				"flows": {
					"implicit": {
						"authorizationUrl": "http://swagger.io/api/oauth/dialog",
						"scopes": {
							"write:pets": "modify things",
							"read:pets": "read things"
						}
					}
				}
			}
		}
	},
	"info": {
		"title": "luminevent",
		"version": "1.2.1",
		"description": "Luminevent is a backend service for managing and organizing events, providing APIs for event creation, user management, and scheduling.",
		"license": {
			"name": "Apache-2.0"
		},
		"contact": {
			"name": "BITS"
		}
	},
	"paths": {
...
		"/events/{id}/seating": {
			"post": {
				"operationId": "CreateSeating",
				"responses": {
					"201": {
						"description": "Seating created successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_SeatingResponse-or-null_"
								}
							}
						}
					}
				},
				"summary": "Creates a new seating arrangement for a specific event.",
				"tags": [
					"Event-Seating"
				],
				"security": [
					{
						"bearer": [
							"admin",
							"staff",
							"hoster"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique ID of the event to which the seating will be added.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The payload containing the seating type and arrangement details.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateSeatingInputBody",
								"description": "The payload containing the seating type and arrangement details."
							}
						}
					}
				}
			},
			"put": {
				"operationId": "UpdateSeating",
				"responses": {
					"200": {
						"description": "Seating updated successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_SeatingResponse-or-null_"
								}
							}
						}
					}
				},
				"summary": "Updates an existing seating arrangement for a specific event.",
				"tags": [
					"Event-Seating"
				],
				"security": [
					{
						"bearer": [
							"admin",
							"staff",
							"hoster"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique ID of the event.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"description": "The payload containing the seating type and fields to be updated.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UpdateSeatingInputBody",
								"description": "The payload containing the seating type and fields to be updated."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetSeating",
				"responses": {
					"200": {
						"description": "Seating fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_SeatingResponse-or-null_"
								}
							}
						}
					}
				},
				"description": "Retrieves the seating arrangement for a specific event.",
				"summary": "Get Seating for an Event",
				"tags": [
					"Event-Seating"
				],
				"security": [
					{
						"bearer": []
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the event.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteSeating",
				"responses": {
					"200": {
						"description": "Seating deleted successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_null_"
								}
							}
						}
					}
				},
				"description": "Deletes the seating arrangement for a specific event.",
				"summary": "Delete Event Seating",
				"tags": [
					"Event-Seating"
				],
				"security": [
					{
						"bearer": [
							"admin",
							"staff",
							"hoster"
						]
					}
				],
				"parameters": [
					{
						"description": "The unique identifier of the event for which to delete the seating.",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/events": {
			"post": {
				"operationId": "CreateEvent",
				"responses": {
					"201": {
						"description": "Created",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_EventOnlyDTO-or-null_"
								}
							}
						}
					}
				},
				"summary": "Creates a new event and its related records.",
				"tags": [
					"Events"
				],
				"security": [
					{
						"bearer": [
							"admin",
							"staff",
							"hoster"
						]
					}
				],
				"parameters": [],
				"requestBody": {
					"description": "The complete event data from the form steps.",
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateEventDTO",
								"description": "The complete event data from the form steps."
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetEvents",
				"responses": {
					"200": {
						"description": "Events fetched successfully",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/APIResponse_PaginatedResult_Events_-or-null_"
								}
							}
						}
					}
				},
				"summary": "Fetches a paginated list of published events.",
				"tags": [
					"Events"
				],
				"security": [],
				"parameters": [
					{
						"description": "The number of events to return per page (defaults to 20).",
						"in": "query",
						"name": "pageSize",
						"required": false,
						"schema": {
							"default": 20,
							"format": "double",
							"type": "number"
						}
					},
					{
						"description": "An optional timestamp string representing the `createdAt` value of the last event from the previous page, used for cursor-based pagination.",
						"in": "query",
						"name": "cursor",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/api/v1"
		}
	]
}